+++ Models/WebSockets/OrderWebSocketModel.cs
@@ -5,6 +5,6 @@ namespace RestaurantSystem.Models.WebSockets
     public class OrderWebSocketModel
     {
         public required WebSocket Socket { get; set; }
-        public required List<int> OrderIds { get; set; }
+        public required List<long> OrderIds { get; set; }
     }
 }
